[NOID] Split out commands file
